# Expose number methods via PyNumberMethods.